requirements.txt
Django==4.1.3
python-telegram-bot==20.7
python-dotenv==1.0.0
gunicorn==21.2.0